requirements.txt
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.103.1
uvicorn==0.23.2
gunicorn==20.1.0